Added local saving path from service location update

+++ app/src/main/java/ru/lobotino/walktraveller/database/model/Point.kt
@@ -6,6 +6,6 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "points")
 data class Point(
     @PrimaryKey(autoGenerate = true) val id: Long = 0,
-    val latitude: Long,
-    val longitude: Long,
+    val latitude: Double,
+    val longitude: Double,
 )
