+++ app/src/main/java/ru/lobotino/walktraveller/model/MapPoint.kt
@@ -1,3 +1,3 @@
 package ru.lobotino.walktraveller.model
 
-data class MapPoint(val latitude: Long, val longitude: Long)
+data class MapPoint(val latitude: Double, val longitude: Double)
